feat: update Scheduler modals to MantineModal
