+++ packages/frontend/src/features/scheduler/components/SchedulerDeleteModal.tsx
@@ -1,22 +1,15 @@
-import {
-    Box,
-    Button,
-    getDefaultZIndex,
-    Group,
-    Loader,
-    Modal,
-    Stack,
-    Text,
-    type ModalProps,
-} from '@mantine-8/core';
+import { Button, getDefaultZIndex, Loader, Stack, Text } from '@mantine-8/core';
 import { IconTrash } from '@tabler/icons-react';
 import { useCallback, useEffect, type FC } from 'react';
 import ErrorState from '../../../components/common/ErrorState';
-import MantineIcon from '../../../components/common/MantineIcon';
+import MantineModal, {
+    type MantineModalProps,
+} from '../../../components/common/MantineModal';
 import { useScheduler } from '../hooks/useScheduler';
 import { useSchedulersDeleteMutation } from '../hooks/useSchedulersDeleteMutation';
 
-interface SchedulerDeleteModalProps extends ModalProps {
+interface SchedulerDeleteModalProps
+    extends Pick<MantineModalProps, 'opened' | 'onClose'> {
     schedulerUuid: string;
     onConfirm: () => void;
 }
@@ -41,51 +34,41 @@ export const SchedulerDeleteModal: FC<SchedulerDeleteModalProps> = ({
     }, [mutation, schedulerUuid]);
 
     return (
-        <Modal
+        <MantineModal
             opened={opened}
-            zIndex={getDefaultZIndex('popover')}
-            title={
-                <Group gap="xs">
-                    <MantineIcon icon={IconTrash} size="lg" color="red" />
-                    <Text fw={600}>Delete scheduled delivery</Text>
-                </Group>
-            }
             onClose={onClose}
-        >
-            <Stack gap="lg">
-                <Box>
-                    {scheduler.isInitialLoading ? (
-                        <Stack h={300} w="100%" align="center">
-                            <Text fw={600}>Loading scheduler</Text>
-                            <Loader />
-                        </Stack>
-                    ) : scheduler.isError ? (
-                        <ErrorState error={scheduler.error.error} />
-                    ) : (
-                        <Text>
-                            Are you sure you want to delete{' '}
-                            <Text fw={700} span>
-                                "{scheduler.data?.name}"
-                            </Text>
-                            ?
-                        </Text>
-                    )}
-                </Box>
-                <Group justify="flex-end" gap="sm">
-                    <Button onClick={onClose} variant="default">
-                        Cancel
+            modalRootProps={{ zIndex: getDefaultZIndex('popover') }}
+            title="Delete scheduled delivery"
+            icon={IconTrash}
+            size="md"
+            actions={
+                scheduler.isSuccess ? (
+                    <Button
+                        loading={mutation.isLoading}
+                        onClick={handleConfirm}
+                        color="red"
+                    >
+                        Delete
                     </Button>
-                    {scheduler.isSuccess && (
-                        <Button
-                            loading={mutation.isLoading}
-                            onClick={handleConfirm}
-                            color="red"
-                        >
-                            Delete
-                        </Button>
-                    )}
-                </Group>
-            </Stack>
-        </Modal>
+                ) : undefined
+            }
+        >
+            {scheduler.isInitialLoading ? (
+                <Stack h={200} w="100%" align="center" justify="center">
+                    <Text fw={600}>Loading scheduler</Text>
+                    <Loader />
+                </Stack>
+            ) : scheduler.isError ? (
+                <ErrorState error={scheduler.error.error} />
+            ) : (
+                <Text>
+                    Are you sure you want to delete{' '}
+                    <Text fw={700} span>
+                        "{scheduler.data?.name}"
+                    </Text>
+                    ?
+                </Text>
+            )}
+        </MantineModal>
     );
 };
