+++ packages/frontend/src/features/scheduler/components/SchedulerModal.tsx
@@ -1,18 +1,33 @@
-import { type ItemsMap } from '@lightdash/common';
-import { Group, Modal, Text } from '@mantine/core';
+import {
+    type ApiError,
+    type ItemsMap,
+    type SchedulerAndTargets,
+} from '@lightdash/common';
+import { Box, Button, Group } from '@mantine-8/core';
 import { IconBell, IconSend } from '@tabler/icons-react';
-import React, { type FC } from 'react';
+import { type UseQueryResult } from '@tanstack/react-query';
+import React, { useState, type FC } from 'react';
 import DocumentationHelpButton from '../../../components/DocumentationHelpButton';
-import MantineIcon from '../../../components/common/MantineIcon';
-import SchedulerModalContent from './SchedulerModalContent';
+import MantineModal from '../../../components/common/MantineModal';
+import { States } from '../utils';
+import { SchedulerModalCreateOrEdit } from './SchedulerModalCreateOrEdit';
+import SchedulersList from './SchedulersList';
 
 const SchedulersModal: FC<
-    Omit<React.ComponentProps<typeof SchedulerModalContent>, 'onClose'> & {
+    Pick<
+        React.ComponentProps<typeof SchedulerModalCreateOrEdit>,
+        | 'resourceUuid'
+        | 'createMutation'
+        | 'isChart'
+        | 'currentParameterValues'
+        | 'availableParameters'
+    > & {
         name: string;
         onClose?: () => void;
         isOpen?: boolean;
         isThresholdAlert?: boolean;
         itemsMap?: ItemsMap;
+        schedulersQuery: UseQueryResult<SchedulerAndTargets[], ApiError>;
     }
 > = ({
     resourceUuid,
@@ -26,61 +41,88 @@ const SchedulersModal: FC<
     availableParameters,
     onClose = () => {},
 }) => {
-    return (
-        <Modal
-            opened={isOpen}
-            onClose={onClose}
-            size="xl"
-            yOffset={65}
-            title={
-                isThresholdAlert ? (
-                    <Group spacing="xs">
-                        <MantineIcon
-                            icon={IconBell}
-                            size="lg"
-                            color="ldGray.7"
-                        />
-                        <Text fw={600}>Alerts</Text>
+    const [modalState, setModalState] = useState<States>(States.LIST);
+    const [schedulerUuidToEdit, setSchedulerUuidToEdit] = useState<
+        string | undefined
+    >();
+    const Actions = () => {
+        if (modalState === States.LIST) {
+            return (
+                <Group>
+                    <Button variant="default" onClick={onClose}>
+                        Cancel
+                    </Button>
+                    <Button onClick={() => setModalState(States.CREATE)}>
+                        Create new
+                    </Button>
+                </Group>
+            );
+        }
+
+        return null;
+    };
+
+    if (modalState === States.LIST) {
+        return (
+            <MantineModal
+                opened={isOpen}
+                onClose={onClose}
+                size="xl"
+                title={isThresholdAlert ? 'Alerts' : 'Scheduled deliveries'}
+                icon={isThresholdAlert ? IconBell : IconSend}
+                headerActions={
+                    isThresholdAlert ? (
                         <DocumentationHelpButton
                             href="https://docs.lightdash.com/guides/how-to-create-alerts"
                             pos="relative"
                             top="2px"
                         />
-                    </Group>
-                ) : (
-                    <Group spacing="xs">
-                        <MantineIcon
-                            icon={IconSend}
-                            size="lg"
-                            color="ldGray.7"
-                        />
-                        <Text fw={600}>Scheduled deliveries</Text>
+                    ) : (
                         <DocumentationHelpButton
                             href="https://docs.lightdash.com/guides/how-to-create-scheduled-deliveries"
                             pos="relative"
                             top="2px"
                         />
-                    </Group>
-                )
-            }
-            styles={(theme) => ({
-                header: { borderBottom: `1px solid ${theme.colors.ldGray[4]}` },
-                body: { padding: 0 },
-            })}
-        >
-            <SchedulerModalContent
+                    )
+                }
+                modalBodyProps={{ bg: 'background' }}
+                actions={<Actions />}
+                cancelLabel={false}
+            >
+                <Box mih={220}>
+                    <SchedulersList
+                        schedulersQuery={schedulersQuery}
+                        isThresholdAlertList={isThresholdAlert}
+                        onEdit={(schedulerUuid) => {
+                            setModalState(States.EDIT);
+                            setSchedulerUuidToEdit(schedulerUuid);
+                        }}
+                    />
+                </Box>
+            </MantineModal>
+        );
+    }
+
+    if (modalState === States.EDIT || modalState === States.CREATE) {
+        return (
+            <SchedulerModalCreateOrEdit
                 resourceUuid={resourceUuid}
-                schedulersQuery={schedulersQuery}
+                schedulerUuidToEdit={
+                    modalState === States.EDIT ? schedulerUuidToEdit : undefined
+                }
                 createMutation={createMutation}
                 onClose={onClose}
+                onBack={() => setModalState(States.LIST)}
                 isChart={isChart}
                 isThresholdAlert={isThresholdAlert}
                 itemsMap={itemsMap}
                 currentParameterValues={currentParameterValues}
                 availableParameters={availableParameters}
             />
-        </Modal>
-    );
+        );
+    }
+
+    return null;
 };
 
 export default SchedulersModal;
