+++ packages/frontend/src/features/scheduler/components/types.ts
@@ -8,3 +8,10 @@ export enum Values {
     FORMATTED = 'formatted',
     RAW = 'raw',
 }
+
+export enum SlackStates {
+    LOADING,
+    SUCCESS,
+    NO_SLACK,
+    MISSING_SCOPES,
+}
